picker/requirements.txt
# Minimal dependencies for picker development
Pillow>=9.0.0
python-dateutil
